requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.7.7
annotated-types==0.7.0
anyio==3.6.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asyncpg==0.29.0
atomicwrites==1.4.1
attrs==21.4.0
bcrypt==3.2.2
cachetools==5.3.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
cryptography==37.0.2
dnspython==2.2.1
email-validator==1.2.1
Faker==12.0.1
fastapi==0.78.0
fastapi-cache2==0.2.1
fastapi-cli==0.0.4
fastapi-users==10.0.4
fastapi-users-db-sqlalchemy==4.0.3
flake8==4.0.1
freezegun==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.13.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.4
makefun==1.13.1
Mako==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mccabe==0.6.1
mdurl==0.1.2
mixer==7.2.2
multidict==6.0.5
packaging==21.3
passlib==1.7.4
pendulum==3.0.0
pluggy==1.5.0
protobuf==4.25.1
pwdlib==0.2.0
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.8
pydantic_core==2.20.1
pyflakes==2.4.0
Pygments==2.18.0
PyJWT==2.3.0
pyparsing==3.0.9
pytest==6.2.5
pytest-asyncio==0.18.3
pytest-freezegun==0.4.2
pytest-pythonpath==0.7.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
time-machine==2.14.2
toml==0.10.2
typer==0.12.3
typing_extensions==4.6.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.17.6
watchfiles==0.22.0
watchgod==0.8.2
websockets==10.3
gunicorn==21.2.0
yarl==1.9.4
redis==4.5.5
slowapi==0.1.8
